+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `plom-scan status` now displays information about bundles and unknown pages.
 * `plom-finish audit` produces a JSON file for post-grading checking/followup/etc
 * Annotator now has a crop tool (in menu) that allows user to excluded unwanted parts of page. Works with undo/rego.
-
+* Annotator now has a "view previous" function that pops up a window that will show previously marked paper(s). Has simple keyboard shortcuts.
 
 ### Changed
 * Plom now requires Python 3.7.